+++ config/services.php
@@ -2,6 +2,10 @@
 
 return [
 
+    'pagespeed' => [
+        'key' => env('PAGESPEED_API_KEY'),
+    ],
+
     /*
     |--------------------------------------------------------------------------
     | Third Party Services